Se modifican relaciones para poder hacer CRUDS

+++ src/main/java/com/TicoTaco/TicoTaco/model/UnidadMedidaModel.java
@@ -11,7 +11,7 @@ public class UnidadMedidaModel {
 
     @Id
     @Column(name = "C_MEDIDA")
-    private int medida;
+    private Long medida;
 
     @Column(name = "D_ABREVIATURA")
     private String abreviatura;
